# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Using a translation service as an example",
        "title": "Go Clean Template API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/proposal/list": {
            "get": {
                "description": "Get a list of proposals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposal"
                ],
                "summary": "Proposal",
                "operationId": "list-proposal",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Proposal"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Customer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "Example Co."
                },
                "segment": {
                    "type": "string",
                    "example": "Enterprise"
                },
                "vertical": {
                    "type": "string",
                    "example": "Banking and Securities"
                }
            }
        },
        "entity.CustomerContact": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/entity.Customer"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+55 (31) 9999 9999"
                }
            }
        },
        "entity.Offering": {
            "type": "object",
            "properties": {
                "addons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Offering"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "Kubernetes Adminstration Training"
                },
                "quantity": {
                    "type": "integer",
                    "example": 12
                },
                "sku": {
                    "type": "string",
                    "example": "KUB101"
                },
                "unitValue": {
                    "type": "number",
                    "example": 4100.5
                }
            }
        },
        "entity.Proposal": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2022-09-15T21:54:42.123Z"
                },
                "customerContact": {
                    "$ref": "#/definitions/entity.CustomerContact"
                },
                "id": {
                    "type": "integer"
                },
                "offerings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Offering"
                    }
                },
                "stage": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}